+++ SnapAudienceNetwork/src/main/java/com/mopub/mobileads/SnapAdBanner.java
@@ -3,6 +3,7 @@ package com.mopub.mobileads;
 import android.app.Activity;
 import android.content.Context;
 import android.text.TextUtils;
+import android.view.View;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -10,45 +11,41 @@ import androidx.annotation.Nullable;
 import com.mopub.common.LifecycleListener;
 import com.mopub.common.Preconditions;
 import com.mopub.common.logging.MoPubLog;
-import com.snap.adkit.dagger.AdKitApplication;
+import com.mopub.common.util.Views;
+import com.snap.adkit.external.BannerView;
 import com.snap.adkit.external.SnapAdClicked;
 import com.snap.adkit.external.SnapAdDismissed;
 import com.snap.adkit.external.SnapAdEventListener;
-import com.snap.adkit.external.SnapAdImpressionHappened;
-import com.snap.adkit.external.SnapAdKit;
 import com.snap.adkit.external.SnapAdKitEvent;
 import com.snap.adkit.external.SnapAdLoadFailed;
 import com.snap.adkit.external.SnapAdLoadSucceeded;
+import com.snap.adkit.external.SnapAdSize;
 import com.snap.adkit.external.SnapAdVisible;
+import com.snap.adkit.external.SnapBannerAdImpressionRecorded;
 
 import java.util.Map;
 
+import static com.mopub.common.DataKeys.ADUNIT_FORMAT;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CLICKED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.DID_DISAPPEAR;
-import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_ATTEMPTED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_FAILED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_SUCCESS;
-import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_ATTEMPTED;
-import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_FAILED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_SUCCESS;
-import static com.mopub.mobileads.MoPubErrorCode.FULLSCREEN_LOAD_ERROR;
-import static com.mopub.mobileads.MoPubErrorCode.NETWORK_NO_FILL;
+import static com.mopub.mobileads.MoPubErrorCode.INLINE_LOAD_ERROR;
 
-public class SnapAdInterstitial extends BaseAd {
-    private static final String ADAPTER_NAME = SnapAdInterstitial.class.getSimpleName();
+public class SnapAdBanner extends BaseAd {
+    private static final String ADAPTER_NAME = SnapAdBanner.class.getSimpleName();
     private static final String SLOT_ID_KEY = "slotId";
 
     private static String mSlotId;
+    private BannerView mBannerView;
 
     private final SnapAdAdapterConfiguration mSnapAdAdapterConfiguration;
 
-    public SnapAdInterstitial() {
+    public SnapAdBanner() {
         mSnapAdAdapterConfiguration = new SnapAdAdapterConfiguration();
     }
-
-    @NonNull
-    private final SnapAdKit snapAdKit = AdKitApplication.getSnapAdKit();
 
     @Nullable
     @Override
@@ -63,7 +60,7 @@ public class SnapAdInterstitial extends BaseAd {
     }
 
     @Override
-    protected void load(@NonNull Context context, @NonNull AdData adData) {
+    protected void load(@NonNull Context context, @NonNull AdData adData) throws Exception {
         Preconditions.checkNotNull(context);
         Preconditions.checkNotNull(adData);
 
@@ -75,18 +72,48 @@ public class SnapAdInterstitial extends BaseAd {
             MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
                     MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR.getIntCode(),
                     MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+
+            if (mLoadListener != null) {
+                mLoadListener.onAdLoadFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+            }
+
+            return;
+        }
+
+        mSnapAdAdapterConfiguration.setCachedInitializationParameters(context, extras);
+
+        String adUnitFormat = extras.get(ADUNIT_FORMAT);
+
+        if (!TextUtils.isEmpty(adUnitFormat)) {
+            adUnitFormat = adUnitFormat.toLowerCase();
+        }
+
+        final SnapAdSize adSize = getAdSize(adUnitFormat);
+
+        if (adSize == null) {
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME,
+                    "SnapAudienceNetwork only supports ad sizes 320*50 and 300*250. " +
+                            "Please ensure your MoPub ad unit format is Banner or Medium Rectangle.");
+            MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
+                    MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR.getIntCode(),
+                    MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+
             if (mLoadListener != null) {
                 mLoadListener.onAdLoadFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
             }
             return;
         }
 
+        mBannerView = new BannerView(context);
+        mBannerView.setAdSize(adSize);
+
         mSlotId = extras.get(SLOT_ID_KEY);
+
         if (!TextUtils.isEmpty(mSlotId)) {
-            snapAdKit.updateSlotId(mSlotId);
+            mBannerView.updateSlotId(mSlotId);
         }
 
-        snapAdKit.setupListener(new SnapAdEventListener() {
+        mBannerView.setupListener(new SnapAdEventListener() {
             @Override
             public void onEvent(SnapAdKitEvent snapAdKitEvent, String slotId) {
                 if (snapAdKitEvent instanceof SnapAdLoadSucceeded) {
@@ -96,11 +123,11 @@ public class SnapAdInterstitial extends BaseAd {
                         mLoadListener.onAdLoaded();
                     }
                 } else if (snapAdKitEvent instanceof SnapAdLoadFailed) {
-                    MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME, FULLSCREEN_LOAD_ERROR.getIntCode(),
-                            FULLSCREEN_LOAD_ERROR);
+                    MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME, INLINE_LOAD_ERROR.getIntCode(),
+                            INLINE_LOAD_ERROR);
 
                     if (mLoadListener != null) {
-                        mLoadListener.onAdLoadFailed(FULLSCREEN_LOAD_ERROR);
+                        mLoadListener.onAdLoadFailed(INLINE_LOAD_ERROR);
                     }
                 } else if (snapAdKitEvent instanceof SnapAdVisible) {
                     MoPubLog.log(getAdNetworkId(), SHOW_SUCCESS, ADAPTER_NAME);
@@ -114,9 +141,9 @@ public class SnapAdInterstitial extends BaseAd {
                     if (mInteractionListener != null) {
                         mInteractionListener.onAdClicked();
                     }
-                } else if (snapAdKitEvent instanceof SnapAdImpressionHappened) {
-                    MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Snap recorded impression: " +
-                            snapAdKitEvent.toString());
+                } else if (snapAdKitEvent instanceof SnapBannerAdImpressionRecorded) {
+                    MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Snap recorded " +
+                            "impression: " + snapAdKitEvent.toString());
 
                     if (mInteractionListener != null) {
                         mInteractionListener.onAdImpression();
@@ -128,40 +155,41 @@ public class SnapAdInterstitial extends BaseAd {
                         mInteractionListener.onAdDismissed();
                     }
                 } else {
-                    MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Received event from Snap " +
-                            "Ad Kit: " + snapAdKitEvent.toString());
+                    MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Received event " +
+                            "from Snap Ad Kit: " + snapAdKitEvent.toString());
                 }
             }
         });
 
-        mSnapAdAdapterConfiguration.setCachedInitializationParameters(context, extras);
-        MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
-
-        snapAdKit.loadInterstitial();
+        mBannerView.loadAd();
     }
 
-    @Override
-    protected void show() {
-        try {
-            MoPubLog.log(getAdNetworkId(), SHOW_ATTEMPTED, ADAPTER_NAME);
-
-            snapAdKit.playAd();
-        } catch (Exception exception) {
-            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Failed to show Snap " +
-                    "Audience Network Ads");
-            MoPubLog.log(getAdNetworkId(), SHOW_FAILED, ADAPTER_NAME,
-                    MoPubErrorCode.NETWORK_NO_FILL.getIntCode(),
-                    MoPubErrorCode.NETWORK_NO_FILL);
-
-            if (mInteractionListener != null) {
-                mInteractionListener.onAdFailed((NETWORK_NO_FILL));
-            }
+    private SnapAdSize getAdSize(String adUnitFormat) {
+        if ("banner".equals(adUnitFormat)) {
+            return SnapAdSize.BANNER;
+        } else if ("medium_rectangle".equals(adUnitFormat)) {
+            return SnapAdSize.MEDIUM_RECTANGLE;
+        } else {
+            return null;
         }
     }
 
     @Override
     protected void onInvalidate() {
-        // no-op
+        Views.removeFromParent(mBannerView);
+
+        if (mBannerView != null) {
+            mBannerView.setupListener(null);
+            mBannerView.destroy();
+
+            mBannerView = null;
+        }
+    }
+
+    @Nullable
+    @Override
+    protected View getAdView() {
+        return mBannerView;
     }
 
     @NonNull